+++ tt/transformations/bexpr.py
@@ -16,6 +16,33 @@ def _ensure_bexpr(expr):
             'argument')
 
 
+def coalesce_negations(expr):
+    """Convert an expression to a form with all negations condensed.
+
+    :returns: A new expression object, transformed so that all "runs" of
+        logical *nots* are condensed into the minimal equivalent number.
+    :rtype: :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
+
+    :raises InvalidArgumentTypeError: If ``expr`` is not a valid type.
+
+    .. code-block:: python
+
+    Here's a simple example showing the basic premise of this transformation::
+
+        >>> from tt import coalesce_negations
+        >>> coalesce_negations('~~A or ~B or ~~~C or ~~~~D')
+        <BooleanExpression "A or ~B or ~C or D">
+
+    This transformation works on expressions, too::
+
+        >>> coalesce_negations('!!(A -> not not B) or ~(~(A xor B))')
+        <BooleanExpression "(A -> B) or (A xor B)">
+
+    """
+    bexpr = _ensure_bexpr(expr)
+    return BooleanExpression(bexpr.tree.root.coalesce_negations())
+
+
 def to_primitives(expr):
     """Convert an expression to a form with only primitive operators.
 
@@ -29,20 +56,21 @@ def to_primitives(expr):
     <tt.expressions.bexpr.BooleanExpression>`
 
     :returns: A new expression object, transformed to contain only primitive
-        operators
+        operators.
     :rtype: :class:`BooleanExpression <tt.expressions.bexpr.BooleanExpression>`
 
     :raises InvalidArgumentTypeError: If ``expr`` is not a valid type.
 
-    .. code-block:: python
+    Here's a simple transformation of exclusive-or::
 
         >>> from tt import to_primitives
         >>> to_primitives('A xor B')
         <BooleanExpression "(A and not B) or (not A and B)">
+
+    And another example of if-and-only-if (using symbolic operators)::
+
         >>> to_primitives('A <-> B')
         <BooleanExpression "(A & B) | (~A & ~B)">
-        >>> to_primitives('(A nand B) -> (C or D)')
-        <BooleanExpression "~(not A or not B) | (C or D)">
 
     """
     bexpr = _ensure_bexpr(expr)